Extract LIKE pattern matching into LikeMatcher
The events Like operator owned the only implementation of the Part 4
LIKE grammar, and alias name search needs the same matching. Move
parsing, matching, and the LRU pattern cache into a reusable
LikeMatcher; the operator delegates with no behavior change. The new
compile() entry point lets callers that evaluate one pattern against
many values parse it once.
